Add support for reading the whole response via a single read call
Add a read method to the ConnectionErrorAwareReadOnlyHTTPResponse
class, which can be called with an optional count argument. If count
is present, read returns at most count bytes. If count is omitted,
read returns the whole response (=> do not use this in case of a large
response body).
Adjust the signature of the ConnectionErrorResponse.read method
accordingly.