Allow an arbitrary encoding in a route specification
Due to security reasons the old code only allowed an ascii encoding (or no
encoding).

Note: our code was _NEVER_ affected by the potential security issue that is
outlined below (since we only allowed an ascii encoding or no encoding at
all).

Doing a regex match on byte level, if the regex was not specified as a bytes
pattern, is wrong and a security issue. For this, consider the following
route:

route('GET', '/xxx/<value>[\xff\x80]{2}', encoding='utf-8').to(
    controller_factory, TestController.repetition
)

From a user's POV this route is supposed to match the following four paths:

/xxx/\xff\xff
/xxx/\xff\x80
/xxx/\x80\xff
/xxx/\x80\x80

[or on byte level (using utf-8 encoding):
 /xxx/\xc3\xbf\xc3\xbf
 /xxx/\xc3\xbf\xc2\x80
 /xxx/\xc2\x80\xc3\xbf
 /xxx/\xc2\x80\xc2\x80]

(that is, if TestController.repetition is called, we expect that the named
match "value" is a str of length 2).

Since the route's path is specified as a str, it is first converted to a
bytes using utf-8 encoding. Using these encoded bytes, a regex is compiled.
Then, this regex is used for matching. Consequently, we would match (among
others) the following path

/xxx/\xc2\xbf

(on byte level)

Since the route matched, TestController.repetition is called and the named
match "value" is a str of length 1 (the inverted question mark (chr(191))).

The fix for this is pretty simple:
If the route's path is specified as a str, the regex match has to be
performed on str level.
If the route's path is specified as a bytes, the regex match has to be
performed on bytes level.