Refactor SelectServer implementation a bit
Now, it it possible to easily select on "arbitrary" fds. In case of
the SelectServer class, Server._accept is only called if there really
is a pending connection on the server socket.
This also fixes a "bug" (well its rather an unexpected behavior because
it still satisfied the documented periodic semantics) in the
PeriodicServer: if a periodic was due and at the "same" time there are
various pending connections, then the periodic is delayed until all
pending connections were accepted. Now, the periodic is executed (if
it is due) before a pending connection is accepted.