Add support for suppressing a connection/read error in a response
This is quite convenient if one only cares if the http
interaction/response processing worked or not. For instance, it can
be used like this

with post(...) as resp:
    f_1(resp)
    ...
    f_n(resp)

print('worked?', not resp.is_connection_error())
# print('worked?', not resp.is_connection_error() and resp.is_success())

The advantage is that one does not have to write a single line of
exception handling.

Note: the current implementation is more or less an ad-hoc
implementation (we should revise it in the future (actually, the whole
http module should be revised)).