Make Server classes more generic
The host and port parameters are replaced by a more abstract address
parameter. Additionally, the sock_family and sock_type parameters
are introduced in order to support different address/protocol families
(for instance, socket.AF_UNIX). Note: the sock_ prefix is used because
without the prefix "sock_type" would be "type" and "type" is a builtin.