Add early close support of the client sock to the router module
Just to avoid unnecessary waste of resources. For this, consider
the following excerpt of a controller method:

    def foo(self):
        ...
        self.reply(200, b'xyz is in progress\n')
        ...
        some lengthy xyz computation
        ...
        return

Once self.reply is called like this, we do not intend to reply any
other data to the client (in fact, the underlying sio is closed).
Even though the sio is closed, the socket remains open until the
computation is done (it is finally closed in Server._accept).
With early_close=True, the socket is closed as soon as its associated
sio is closed.