Add example that works with a CMSG of the type TCP_CM_INQ
Its sole purpose is to demonstrate how to use our framework in order
to specify a CMSG processor etc. for a specific CMSG type. In particular,
we will setup a controller that receives after each recvmsg (syscall)
the number of remaining bytes in the TCP receive queue. For this, we
have to setup a CMSG processor that parses a CMSG of the type
TCP_CM_INQ.